backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
scikit-learn
numpy
